Ecs added (WIP)
to render multiple objects, I can create multiple entities from world and give them a model and transform component. The render pass will then iterate through the entites and use those values for draw calls. Still a work in progress since I might refactor how the model/mesh drawing code. The transform struct also still needs to implement rotation, and I could probably optimise the matrix construction for it (the implementation right now is to construct a matrix for scale and translation and then to multiply. Im not sure how expensive it is for cgmath to construct a mat4, but if it is, I would rather just look into the math and construct it manually)