Seperate model from texture loading
Very temporary for now, but just wanted to make the change so that I could give any model any texture and vice versa. Will soon refactor the texture loading to make it nicer and incorporate it into the ECS system so I don't have to pass bind groups for every created texture.