glTF loading (WIP)
very temporary for now. No animation support or anything crazy, just grabs vertex and index info and creates a model, same as the .obj loader. added temporary fix to scene.rs to support loading .obj and gltf, but it's not good. If the functions for loading obj and gltf still return the same type in the future, I'll alter it so that they use the same value in the .json (model_path) and i'll just check the end of the string to decide wether it should use load_model or load_gltf